+++ fly/consumer-tx/fly.toml
@@ -21,7 +21,7 @@ kill_timeout = "5s"
   internal_port = 0
   auto_stop_machines = false
   auto_start_machines = true
-  min_machines_running = 16
+  min_machines_running = 3
 
 [checks]
   [checks.health]
